Сделаны уведомления и переделан адаптер в HomeFragment и ProfileFragment. Сделано отображение колва лайков, историй и имени профиля.

+++ app/src/main/java/dem/corp/story/story/StoryAdapter.kt
@@ -1,12 +1,15 @@
 package dem.corp.story.story
 
 import android.annotation.SuppressLint
+import android.app.Dialog
+import android.content.Context
 import android.content.Intent
+import android.graphics.Color
+import android.graphics.drawable.ColorDrawable
 import android.os.Build
+import android.provider.Contacts
 import android.util.Log
-import android.view.LayoutInflater
-import android.view.View
-import android.view.ViewGroup
+import android.view.*
 import android.widget.ImageButton
 import android.widget.ImageView
 import android.widget.TextView
@@ -22,6 +25,7 @@ import com.google.firebase.database.FirebaseDatabase
 import com.google.firebase.database.ValueEventListener
 import dem.corp.story.CommentActivity
 import dem.corp.story.R
+import dem.corp.story.models.Comment
 import dem.corp.story.models.Story
 import dem.corp.story.repository.firebase.*
 import java.text.SimpleDateFormat
@@ -31,29 +35,34 @@ import java.time.format.DateTimeFormatter
 import java.util.*
 import kotlin.collections.HashMap
 
-internal class StoryAdapter(options: FirebaseRecyclerOptions<Story>) :
-    FirebaseRecyclerAdapter<Story, dem.corp.story.story.StoryAdapter.viewholder>(options) {
+class StoryAdapter internal constructor(context: Context?, stories: List<Story>, isProfileFr: Boolean) :
+    RecyclerView.Adapter<StoryAdapter.ViewHolder>() {
+    private val inflater: LayoutInflater
+    private val stories: List<Story>
+    private val isProfileFr: Boolean
+//    private val isProfileFragment: Boolean
 
-    override fun getItem(position: Int): Story {
-        return super.getItem(itemCount - 1 - position)
+    override fun onCreateViewHolder(parent: ViewGroup, viewType: Int): ViewHolder {
+        var view: View? = null
+        if (isProfileFr){
+            view = inflater.inflate(R.layout.story_item_profile, parent, false)
+
+        }
+        else {
+            view = inflater.inflate(R.layout.story_item, parent, false)
+        }
+        return ViewHolder(view)
     }
 
-
     @RequiresApi(Build.VERSION_CODES.O)
-    @SuppressLint("SimpleDateFormat")
-    override fun onBindViewHolder(
-        holder: StoryAdapter.viewholder,
-        i: Int,
-        post: Story
-    ) {
-        var i = i
-        i = itemCount - 1 - i
-
-        holder.title.text = post.title
-        holder.date.text = if (post.date != "") {
+    override fun onBindViewHolder(holder: ViewHolder, position: Int) {
+        var i2 = itemCount - 1 - position
+        val model: Story = stories[i2]
+        holder.title.text = model.title
+        holder.date.text = if (model.date != "" && model.date != null) {
             val firstFormatter = DateTimeFormatter.ofPattern("dd.MM.yyyy hh:mm")
             var period = Period.between(
-                LocalDate.parse(post.date, firstFormatter),
+                LocalDate.parse(model.date, firstFormatter),
                 LocalDate.parse(
                     SimpleDateFormat("dd.MM.yyyy hh:mm").format(Date()),
                     firstFormatter
@@ -64,25 +73,38 @@ internal class StoryAdapter(options: FirebaseRecyclerOptions<Story>) :
             else "today"
         } else "today"
 
-        val finalI = i
-
         holder.commentBtn.setOnClickListener { v ->
             val intent = Intent(v.context, CommentActivity::class.java)
-            intent.putExtra("storyTitle", post.title)
-            intent.putExtra("storyID", getRef(finalI).key)
+            intent.putExtra("storyTitle", model.title)
+            intent.putExtra("storyID", model.id)
+            intent.putExtra("storyFrom", model.from)
             v.context.startActivity(intent)
         }
 
         holder.updateFields()
 
         holder.read.setOnClickListener { v ->
-            val bottomSheetDialog = RoundedBottomSheetDialog(v.context)
-            bottomSheetDialog.setContentView(R.layout.read_story_bottom_sheet)
-            bottomSheetDialog.show()
-            val storyText = bottomSheetDialog.findViewById<TextView>(R.id.story_txt)
-            val storyTitle = bottomSheetDialog.findViewById<TextView>(R.id.story_title)
-            var storyDate = bottomSheetDialog.findViewById<TextView>(R.id.tv_btm_shed_date)
-            val key = getRef(finalI).key
+            val dialog = Dialog(v.context)
+            dialog.requestWindowFeature(Window.FEATURE_NO_TITLE)
+            dialog.setContentView(R.layout.read_story_bottom_sheet)
+
+            val storyText = dialog.findViewById<TextView>(R.id.story_txt)
+            val storyTitle = dialog.findViewById<TextView>(R.id.story_title)
+            val storyDate = dialog.findViewById<TextView>(R.id.tv_btm_shed_date)
+            val storyFrom = dialog.findViewById<TextView>(R.id.story_from)
+            val storyClose = dialog.findViewById<ImageButton>(R.id.dialog_close)
+
+            dialog.show()
+            dialog.window?.setLayout(ViewGroup.LayoutParams.MATCH_PARENT,ViewGroup.LayoutParams.WRAP_CONTENT)
+            dialog.window?.setBackgroundDrawable(ColorDrawable(Color.TRANSPARENT))
+            dialog.window?.attributes?.windowAnimations = R.style.DialogAnimation
+            dialog.window?.setGravity(Gravity.BOTTOM)
+
+            storyClose.setOnClickListener {
+                dialog.dismiss()
+            }
+
+            val key = model.id.toString()
             var e = true
             FirebaseDatabase.getInstance().getReference("Stories").child(key!!)
                 .addValueEventListener(object : ValueEventListener {
@@ -94,10 +116,10 @@ internal class StoryAdapter(options: FirebaseRecyclerOptions<Story>) :
                             storyText!!.text = text
                             storyTitle!!.text = title
                             Log.d("TAG", date)
-                            storyDate!!.text = if (post.date != "") {
+                            storyDate!!.text = if (model.date != "") {
                                 val firstFormatter = DateTimeFormatter.ofPattern("dd.MM.yyyy hh:mm")
                                 var period = Period.between(
-                                    LocalDate.parse(post.date, firstFormatter),
+                                    LocalDate.parse(model.date, firstFormatter),
                                     LocalDate.parse(
                                         SimpleDateFormat("dd.MM.yyyy hh:mm").format(Date()),
                                         firstFormatter
@@ -115,17 +137,11 @@ internal class StoryAdapter(options: FirebaseRecyclerOptions<Story>) :
         }
     }
 
-    override fun onCreateViewHolder(
-        parent: ViewGroup,
-        viewType: Int
-    ): StoryAdapter.viewholder {
-        val view: View = LayoutInflater.from(parent.context)
-            .inflate(R.layout.story_item, parent, false)
-        Log.d("j", "Adapter started")
-        return viewholder(view)
+    override fun getItemCount(): Int {
+        return stories.size
     }
 
-    internal inner class viewholder(itemView: View) : RecyclerView.ViewHolder(itemView) {
+    inner class ViewHolder(view: View) : RecyclerView.ViewHolder(view) {
         val title: TextView = itemView.findViewById(R.id.story_title)
         val date: TextView = itemView.findViewById(R.id.tv_story_date)
         val read: ImageButton = itemView.findViewById(R.id.read_more)
@@ -147,7 +163,7 @@ internal class StoryAdapter(options: FirebaseRecyclerOptions<Story>) :
 
         private fun updateStoryImage() {
             try {
-                val url = getItem(layoutPosition).imageUrl
+                val url = stories[itemCount-1-layoutPosition].imageUrl
                 if (url != "")
                     Glide.with(itemView).load(url).into(storyImage)
             }catch (e: Exception){
@@ -157,7 +173,7 @@ internal class StoryAdapter(options: FirebaseRecyclerOptions<Story>) :
 
         private fun updateLikeImage() {
             try {
-                if (getItem(layoutPosition).getLikesList().contains(AUTH.uid!!)) {
+                if (stories[itemCount-1-position].getLikesList().contains(AUTH.uid!!)) {
                     setLikeIsClicked()
                 } else {
                     setLikeIsNotClicked()
@@ -169,7 +185,7 @@ internal class StoryAdapter(options: FirebaseRecyclerOptions<Story>) :
 
         private fun updateLikesCountText() {
             try {
-                likesCount.text = getItem(position).likes.keys.size.toString()
+                likesCount.text = stories[itemCount-1-position].likes.keys.size.toString()
             } catch (e: Exception) {
                 e.printStackTrace()
             }
@@ -191,12 +207,17 @@ internal class StoryAdapter(options: FirebaseRecyclerOptions<Story>) :
                                 .addValueEventListener(object : ValueEventListener {
                                     override fun onDataChange(snapshot2: DataSnapshot) {
                                         if (a) {
+                                            val date = Date()
+                                            val formatForDate = SimpleDateFormat("dd.MM.yyyy hh:mm")
                                             val from = snapshot2.getValue().toString()
                                             val map = HashMap<String, String>()
                                             map.put("from", from)
                                             map.put("username", username)
                                             map.put("storyID", key)
                                             map.put("type", "like")
+                                            map.put("date", formatForDate.format(date))
+                                            map.put("text", "like")
+
                                             DATABASE_ROOT.child("Users").child(UID)
                                                 .child("notifications").child(key).setValue(map)
                                             a = false
@@ -211,25 +232,6 @@ internal class StoryAdapter(options: FirebaseRecyclerOptions<Story>) :
 
                     override fun onCancelled(error: DatabaseError) {}
                 })
-//            DATABASE_ROOT.child(NODE_USERS).child(UID).child(CHILD_USERNAME).addValueEventListener(ValueEventListener)
-//            DATABASE_ROOT.child(NODE_USERS).child(UID).child(CHILD_USERNAME).get().addOnCompleteListener { task ->
-//                if (task.isSuccessful) {
-//                    val username = task.result.toString()
-//                    DATABASE_ROOT.child(NODE_USERS).child(UID).child(CHILD_USERNAME).get().addOnCompleteListener { task2 ->
-//                        if (task2.isSuccessful){
-//                            val from = task2.result.toString()
-//                            val map = HashMap<String, String> ()
-//                            map.put("from", from)
-//                            map.put("username", username)
-//                            map.put("storyID", key)
-//                            map.put("type", "like")
-//                            DATABASE_ROOT.child("Users").child(UID).child("notifications").child(key).setValue(map)
-//                        }
-//                    }
-//                } else {
-//                    task.exception?.printStackTrace()
-//                }
-//            }
         }
 
         private fun deleteNotification(position: Int, key: String) {
@@ -238,25 +240,32 @@ internal class StoryAdapter(options: FirebaseRecyclerOptions<Story>) :
 
         private fun initOnClick() {
             likeBtn.setOnClickListener {
-                val story = getItem(position)
-                val key = getRef(itemCount - 1 - position).key.toString()
+                val story = stories[itemCount - 1-position]
+                val key = stories[itemCount-1-position].id.toString()
                 if (story.getLikesList().contains(AUTH.uid!!)) {
                     removeLikeFromStory(story) {
                         story.likes.remove(AUTH.uid!!)
                         setLikeIsNotClicked()
-                        deleteNotification(position, key)
+                        if (!story.from.equals(UID))
+                            deleteNotification(position, key)
                         likesCount.text = story.likes.size.toString()
                     }
                 } else {
                     putLikeToStory(story) {
                         story.likes[AUTH.uid!!] = ""
                         setLikeIsClicked()
-                        sendNotification(position, key)
+                        if (!story.from.equals(UID))
+                            sendNotification(position, key)
                         likesCount.text = story.likes.size.toString()
                     }
                 }
             }
         }
+    }
 
+    init {
+        this.stories = stories
+        this.isProfileFr = isProfileFr
+        inflater = LayoutInflater.from(context)
     }
 }
